+++ webpack.config.js
@@ -63,14 +63,17 @@ module.exports = Object.assign({}, baseConfig.webpackOptions, {
       }
     }),
 
-    // TODO As of webpack 2.1.0-beta.24, disable this to prevent "Template cannot be applied as TemplateArgument: HarmonyImportDependency". See https://github.com/webpack/webpack/issues/2644
-    // Prevents the inclusion of duplicate code into bundle and instead applies a copy
-    // of the function at runtime, which results smaller file size
-    // new webpack.optimize.DedupePlugin(),
-
     // Generates `index.html` at the location specified by the user
     new HtmlWebpackPlugin(Object.assign({}, baseConfig.htmlWebpackPluginOptions, {
       filename: path.join(__dirname, packageJson.config.entry_file_path)
     }))
-  )
+  ),
+
+  // To suppress this warning when creating the vendor bundle:-
+  //
+  // WARNING in asset size limit: The following asset(s) exceed the recommended size limit (250 kB).
+  // This can impact web performance.
+  performance: {
+    hints: false
+  }
 });